requirements.txt
autopep8==1.6.0
click==8.0.3
Flask==2.0.3
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
mysql-connector-python==8.0.28
protobuf==3.19.4
pycodestyle==2.8.0
toml==0.10.2
Werkzeug==2.0.3
